Use categories total instead of counting spent and received separately
There is an issue when in a single category you spend and receive money
at the same time. E.g. when you have a savings account, you send some
money to that account but later you have to take some out. Previously we
counted both transactions separately which was a little confusing. Now
we count the total, which in this example represents how much you were
able to save, even if you made multiple transactions there and back.